Refactored to common module and channel module, Add channel and channellisting flows to channel module

+++ assure/src/main/java/com/increff/assure/dao/OrderDao.java
@@ -23,7 +23,7 @@ public class OrderDao extends AbstractDao {
         em.persist(p);
     }
 
-    public OrderPojo select(int id) {
+    public OrderPojo select(Long id) {
         TypedQuery<OrderPojo> query = getQuery(SELECT_ID, OrderPojo.class);
         query.setParameter("id", id);
         return getSingle(query);
